reduce copy-pasted setup
