+++ packages/react-app-rewired/scripts/test.js
@@ -1,13 +1,3 @@
-// @remove-on-eject-begin
-/**
- * Copyright (c) 2015-present, Facebook, Inc.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
- */
-// @remove-on-eject-end
 "use strict";
 
 // Do this as the first thing so that any code reading it knows the right env.
@@ -22,8 +12,9 @@ process.on("unhandledRejection", err => {
   throw err;
 });
 
+const paths = require("./utils/paths");
 // Ensure environment variables are read.
-require("../config/env");
+require(paths.scriptVersion + "/config/env");
 
 const jest = require("jest");
 const argv = process.argv.slice(2);
@@ -33,12 +24,9 @@ if (!process.env.CI && argv.indexOf("--coverage") < 0) {
   argv.push("--watch");
 }
 
-const paths = require("../config/paths");
 
-// @remove-on-eject-begin
-// This is not necessary after eject because we embed config into package.json.
 const createJestConfig = require("./utils/createJestConfig");
-const override = require(paths.projectDir + '/config-overrides');
+const override = require(paths.configOverrides);
 const overrideFn = (typeof override === 'function' || typeof override.jest !== 'function')
   ? (config) => config
   : override.jest;
@@ -55,5 +43,5 @@ argv.push(
     ))
   )
 );
-// @remove-on-eject-end
+
 jest.run(argv);
