+++ packages/react-app-rewired/scripts/utils/babelTransform.js
